#350: fix and test TranscriptSequenceDecorator#getCodonAt
The `endPos` refers to a half-open interval, i.e. it is not included (as
an index, it would refer to the start of the next codon, if available),
so the sequence length check should be less strict ('<' instead of '<=').

Also added a test for TranscriptSequenceDecorator#getCodonsStartingFrom,
but this is not affected by #350.