Ensure RefSeq parser creates non-empty CDS regions
The problem was that the parser skipped over all CDS records because of
a prior null check, while CDS records have no transcript IDs
(as correctly pointed out in a comment in
RefSeqParser#loadTranscriptModels).

This fix also changes the parser so that an IllegalStateException gets
thrown in case a coding transcript (ie one with an 'NM_*' / 'XM_*'
accession) gets assigned an empty CDS region.

This seems to be the safest behavior in such cases, as it indicates
corrupt or incomplete reference data in the GFF file and would otherwise
fail silently.